apps/backend/pyproject.toml
[project]
name = "bird-backend"
version = "0.1.0"
description = "FastAPI backend for bird detection"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.8.0",
  "pillow>=10.4.0",
  "torch>=2.3.0",
  "torchvision>=0.18.0",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
